Refactored the cover download code into it's own service, where it makes more sense.
