+++ API.Benchmark/ArchiveServiceBenchmark.cs
@@ -32,7 +32,7 @@ public class ArchiveServiceBenchmark
     public ArchiveServiceBenchmark()
     {
         _directoryService = new DirectoryService(null, new FileSystem());
-        _imageService = new ImageService(null, _directoryService, Substitute.For<IEasyCachingProviderFactory>());
+        _imageService = new ImageService(null, _directoryService);
         _archiveService = new ArchiveService(new NullLogger<ArchiveService>(), _directoryService, _imageService, Substitute.For<IMediaErrorService>());
     }
