+++ API.Tests/Services/BookServiceTests.cs
@@ -17,7 +17,7 @@ public class BookServiceTests
     {
         var directoryService = new DirectoryService(Substitute.For<ILogger<DirectoryService>>(), new FileSystem());
         _bookService = new BookService(_logger, directoryService,
-            new ImageService(Substitute.For<ILogger<ImageService>>(), directoryService, Substitute.For<IEasyCachingProviderFactory>())
+            new ImageService(Substitute.For<ILogger<ImageService>>(), directoryService)
             , Substitute.For<IMediaErrorService>());
     }
