+++ API/API.csproj
@@ -106,6 +106,7 @@
     <PackageReference Include="System.IO.Abstractions" Version="21.0.29" />
     <PackageReference Include="System.Drawing.Common" Version="8.0.10" />
     <PackageReference Include="VersOne.Epub" Version="3.3.2" />
+    <PackageReference Include="YamlDotNet" Version="16.1.3" />
   </ItemGroup>
 
   <ItemGroup>
